Merges CpuManager to Vmm
